+++ lib/ex/parser/scanner_command_buffers.py
@@ -3,7 +3,6 @@ from .tokens import TokenEof
 from .tokens_base import TOKEN_COMMAND_BUFFERS
 from .tokens_base import TokenOfCommand
 from NeoVintageous.lib import ex
-from NeoVintageous.lib import nvim
 
 
 @ex.command('buffers', 'buffers')
@@ -19,6 +18,6 @@ def scan_command_buffers(state):
     try:
         state.expect(EOF)
     except ValueError:
-        raise nvim.Error(nvim.E_TRAILING_CHARS)
+        raise Exception("E488: Trailing characters")
 
     return None, [TokenCommandBuffers(), TokenEof()]