+++ lib/ex/parser/scanner_command_read_shell_out.py
@@ -3,7 +3,6 @@ from .tokens import TokenEof
 from .tokens_base import TOKEN_COMMAND_READ_SHELL_OUT
 from .tokens_base import TokenOfCommand
 from NeoVintageous.lib import ex
-from NeoVintageous.lib import nvim
 
 
 plus_plus_translations = {
@@ -52,7 +51,7 @@ def scan_command_read_shell_out(state):
         # https://vimhelp.appspot.com/editing.txt.html#[++opt]
         m = state.expect_match(
             r'(?:f(?:ile)?f(?:ormat)?|(?:file)?enc(?:oding)?|(?:no)?bin(?:ary)?|bad|edit)(?=\s|$)',
-            lambda: nvim.Error(nvim.E_INVALID_ARGUMENT))
+            lambda: Exception("E474: Invalid argument"))
         name = m.group(0)
         params['++'] = plus_plus_translations.get(name, name)
         state.ignore()