+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes are documented in this file using the [Keep a CHANGELOG](htt
 
 ### Fixed
 
+* Fixed: Error message typos and grammar
 * Fixed [#254](https://github.com/NeoVintageous/NeoVintageous/issues/245): `:vs[plit] [file]` TypeError: unsupported operand type(s) for +: 'NoneType' and 'str'
 
 ## Unreleased (1.4.4)